+++ src/ordinary_transaction.rs
@@ -7,7 +7,7 @@
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific TON DEV software governing permissions and
+* See the License for the specific EVERX DEV software governing permissions and
 * limitations under the License.
 */
 
@@ -20,23 +20,16 @@ use std::sync::atomic::AtomicU64;
 use std::sync::atomic::Ordering;
 #[cfg(feature = "timings")]
 use std::time::Instant;
-use ton_block::{
+use ever_block::{
     AccStatusChange, Account, AccountStatus, AddSub, CommonMsgInfo, Grams, Message,
     Serializable, TrBouncePhase, TrComputePhase, Transaction, TransactionDescr,
     TransactionDescrOrdinary, MASTERCHAIN_ID, GlobalCapabilities
 };
-use ton_types::{error, fail, Result, HashmapType, SliceData};
-use ton_vm::{
+use ever_block::{error, fail, Result, HashmapType, SliceData};
+use ever_vm::{
     boolean, int,
     stack::{integer::IntegerData, Stack, StackItem}, SmartContractInfo,
 };
-
-
-
-
-
-
-
 
 pub struct OrdinaryTransactionExecutor {
     config: BlockchainConfig,
@@ -218,7 +211,7 @@ impl TransactionExecutor for OrdinaryTransactionExecutor {
         let config_params = self.config().raw_config().config_params.data().cloned();
         let mut smc_info = SmartContractInfo {
             capabilities: self.config().raw_config().capabilities(),
-            myself: SliceData::load_builder(account_address.write_to_new_cell().unwrap_or_default()).unwrap(),
+            myself: SliceData::load_builder(account_address.write_to_new_cell().unwrap_or_default())?,
             block_lt: params.block_lt,
             trans_lt: lt,
             unix_time: params.block_unixtime,
@@ -394,11 +387,11 @@ impl TransactionExecutor for OrdinaryTransactionExecutor {
     }
     fn ordinary_transaction(&self) -> bool { true }
     fn config(&self) -> &BlockchainConfig { &self.config }
-    fn build_stack(&self, in_msg: Option<&Message>, account: &Account) -> Stack {
+    fn build_stack(&self, in_msg: Option<&Message>, account: &Account) -> Result<Stack> {
         let mut stack = Stack::new();
         let in_msg = match in_msg {
             Some(in_msg) => in_msg,
-            None => return stack
+            None => return Ok(stack)
         };
         let acc_balance = int!(account.balance().map_or(0, |value| value.grams.as_u128()));
         let msg_balance = int!(in_msg.get_value().map_or(0, |value| value.grams.as_u128()));
@@ -411,6 +404,6 @@ impl TransactionExecutor for OrdinaryTransactionExecutor {
             .push(StackItem::Cell(in_msg_cell))
             .push(StackItem::Slice(body_slice))
             .push(function_selector);
-        stack
+        Ok(stack)
     }
 }