+++ src/tick_tock_transaction.rs
@@ -1,5 +1,5 @@
 /*
-* Copyright (C) 2019-2022 TON Labs. All Rights Reserved.
+* Copyright (C) 2019-2023 EverX. All Rights Reserved.
 *
 * Licensed under the SOFTWARE EVALUATION License (the "License"); you may not use
 * this file except in compliance with the License.
@@ -7,19 +7,19 @@
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific TON DEV software governing permissions and
+* See the License for the specific EVERX DEV software governing permissions and
 * limitations under the License.
 */
 
 use crate::{blockchain_config::BlockchainConfig, ExecuteParams, TransactionExecutor, error::ExecutorError, ActionPhaseResult};
 
 use std::sync::atomic::Ordering;
-use ton_block::{
+use ever_block::{
     Account, CurrencyCollection, Grams, Message, TrComputePhase, Transaction, 
     TransactionDescr, TransactionDescrTickTock, TransactionTickTock, Serializable
 };
-use ton_types::{error, fail, Result, HashmapType, SliceData};
-use ton_vm::{
+use ever_block::{error, fail, Result, HashmapType, SliceData};
+use ever_vm::{
     boolean, int,
     stack::{integer::IntegerData, Stack, StackItem}, SmartContractInfo,
 };
@@ -111,7 +111,7 @@ impl TransactionExecutor for TickTockTransactionExecutor {
         let config_params = self.config().raw_config().config_params.data().cloned();
         let mut smc_info = SmartContractInfo {
             capabilities: self.config().raw_config().capabilities(),
-            myself: SliceData::load_builder(account_address.write_to_new_cell().unwrap_or_default()).unwrap(),
+            myself: SliceData::load_builder(account_address.write_to_new_cell().unwrap_or_default())?,
             block_lt: params.block_lt,
             trans_lt: lt,
             unix_time: params.block_unixtime,
@@ -221,16 +221,16 @@ impl TransactionExecutor for TickTockTransactionExecutor {
     }
     fn ordinary_transaction(&self) -> bool { false }
     fn config(&self) -> &BlockchainConfig { &self.config }
-    fn build_stack(&self, _in_msg: Option<&Message>, account: &Account) -> Stack {
-        let account_balance = account.balance().unwrap().grams.as_u128();
-        let account_id = account.get_id().unwrap();
+    fn build_stack(&self, _in_msg: Option<&Message>, account: &Account) -> Result<Stack> {
+        let account_balance = account.balance().ok_or_else(|| failure::format_err!("Can't get account balance."))?.grams.as_u128();
+        let account_id = account.get_id().ok_or_else(|| failure::format_err!("Can't get account id."))?;
         let mut stack = Stack::new();
         stack
             .push(int!(account_balance))
             .push(StackItem::integer(IntegerData::from_unsigned_bytes_be(account_id.get_bytestring(0))))
             .push(boolean!(self.tt.is_tock()))
             .push(int!(-2));
-        stack
+        Ok(stack)
     }
 }