+++ CHANGELOG.md
@@ -2,6 +2,49 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 1.17.0
+
+- the crate was renamed from `ton_executor` to `ever_executor`
+- supported renaming of other crates
+
+## Version 1.16.122
+
+- Do not delete frozen accounts (if related capability set)
+
+## Version 1.16.108
+
+- Deny non-zero cell level in code/data/lib
+
+## Version 1.16.106
+
+- Remove compiler warning
+- Remove unwraps that lead to panic
+
+## Version 1.16.85
+
+- Deny ChangeLibrary action when CapSetLibCode is unset
+
+## Version 1.16.40
+
+- Disable debug symbols by default
+
+## Version 1.16.0
+
+- Skiped compute phase for suspended addresses
+
+## Version 1.15.196
+
+- Removed extra crates bas64
+- Minor refactoring
+
+## Version 1.15.191
+
+- Supported ever-types version 2.0
+
+## Version 1.15.190
+
+- Add test for CapFeeInGasUnits
+
 ## Version: 1.15.188
 
 ### News